Moved partial and full env var matching as static functions in Env
