+++ src/env.rs
@@ -7,6 +7,9 @@ use std::ops::Index;
 
 lazy_static! {
     static ref ENV_VAR_REGEX: Regex = Regex::new(r"(\$[\w\?\-#!\$_@\*]*)").unwrap();
+    static ref PARTIAL_BRACKET_ENV_VAR_REGEX: Regex =
+        Regex::new(r"(\$\{([\w\?\-#!\$_@\*]*)\}?)").unwrap();
+    static ref BRACKET_ENV_VAR_REGEX: Regex = Regex::new(r"(\$\{([\w\?\-#!\$_@\*]+)\})").unwrap();
 }
 
 type Key = String;
@@ -122,6 +125,50 @@ impl Env {
                 .into_owned();
         }
         res
+    }
+
+    // TODO: -> Option<String>
+    /// Returns environment variable at position in text.
+    pub fn var_at_pos(pos: usize, text: &str) -> Value {
+        assert!(pos <= text.len());
+        for cap in BRACKET_ENV_VAR_REGEX.captures_iter(text) {
+            let cap0 = cap.get(0).unwrap();
+            let cap1 = cap.get(1);
+            if pos >= cap0.start() && pos <= cap0.end() {
+                if let Some(cap1_val) = cap1 {
+                    return cap1_val.as_str().to_string();
+                }
+            }
+        }
+        for cap in ENV_VAR_REGEX.captures_iter(text) {
+            let cap = cap.get(0).unwrap();
+            if pos >= cap.start() && pos <= cap.end() {
+                return cap.as_str().to_string();
+            }
+        }
+        "".to_string()
+    }
+
+    // TODO: return Option<Value>
+    /// Returns partial environment variable at position in text.
+    pub fn partial_var_at_pos(pos: usize, text: &str) -> Value {
+        assert!(pos <= text.len());
+        for cap in PARTIAL_BRACKET_ENV_VAR_REGEX.captures_iter(text) {
+            let cap0 = cap.get(0).unwrap();
+            let cap1 = cap.get(1);
+            if pos >= cap0.start() && pos <= cap0.end() {
+                if let Some(cap1_val) = cap1 {
+                    return cap1_val.as_str().to_string();
+                }
+            }
+        }
+        for cap in ENV_VAR_REGEX.captures_iter(text) {
+            let cap = cap.get(0).unwrap();
+            if pos >= cap.start() && pos <= cap.end() {
+                return cap.as_str().to_string();
+            }
+        }
+        "".to_string()
     }
 }
 
@@ -309,5 +356,111 @@ mod tests {
         env.insert("USERNAME".to_string(), "foobar".to_string());
         let output = env.replace_vars(&input);
         assert_eq!(output, "foobar".to_string());
+    }
+
+    #[test]
+    fn partial_env_var_at_pos_start() {
+        assert_eq!(
+            Env::partial_var_at_pos(6, "hello ${world and universe"),
+            "${world"
+        );
+    }
+
+    #[test]
+    fn partial_env_var_at_pos_middle() {
+        assert_eq!(
+            Env::partial_var_at_pos(9, "hello ${world and universe"),
+            "${world"
+        );
+    }
+
+    #[test]
+    fn partial_env_var_at_pos_end() {
+        assert_eq!(
+            Env::partial_var_at_pos(12, "hello ${world and universe"),
+            "${world"
+        );
+    }
+
+    #[test]
+    fn partial_env_var_at_pos_can_yield_full_match() {
+        assert_eq!(
+            Env::partial_var_at_pos(9, "hello ${world} and universe"),
+            "${world}"
+        );
+    }
+
+    #[test]
+    fn partial_env_var_at_pos_only_dollar_sign_bracket() {
+        assert_eq!(Env::partial_var_at_pos(6, "hello ${  and universe"), "${");
+    }
+
+    #[test]
+    fn partial_env_var_at_pos_only_dollar_sign_bracket_dash() {
+        assert_eq!(Env::partial_var_at_pos(6, "hello ${-  and universe"), "${-");
+    }
+
+    #[test]
+    fn env_var_at_pos_beginning() {
+        assert_eq!(Env::var_at_pos(0, "$hello world and universe"), "$hello");
+    }
+
+    #[test]
+    fn env_var_at_pos_start() {
+        assert_eq!(Env::var_at_pos(6, "hello $world and universe"), "$world");
+    }
+
+    #[test]
+    fn env_var_at_pos_middle() {
+        assert_eq!(Env::var_at_pos(2, "$hello world and universe"), "$hello");
+    }
+
+    #[test]
+    fn env_var_at_pos_end() {
+        assert_eq!(Env::var_at_pos(11, "hello $world and universe"), "$world");
+    }
+
+    #[test]
+    fn env_var_at_pos_right_after() {
+        assert_eq!(Env::var_at_pos(12, "hello $world and universe"), "$world");
+    }
+
+    #[test]
+    fn env_var_at_pos_after() {
+        assert_eq!(Env::var_at_pos(13, "hello $world  and universe"), "");
+    }
+
+    #[test]
+    fn env_var_at_pos_only_dollar_sign() {
+        assert_eq!(Env::var_at_pos(6, "hello $  and universe"), "$");
+    }
+
+    #[test]
+    fn env_var_at_pos_dollar_dash() {
+        assert_eq!(Env::var_at_pos(6, "hello $- and universe"), "$-");
+    }
+
+    #[test]
+    fn bracket_env_var_at_pos_start() {
+        assert_eq!(
+            Env::var_at_pos(6, "hello ${world} and universe"),
+            "${world}"
+        );
+    }
+
+    #[test]
+    fn bracket_env_var_at_pos_middle() {
+        assert_eq!(
+            Env::var_at_pos(10, "hello ${world} and universe"),
+            "${world}"
+        );
+    }
+
+    #[test]
+    fn bracket_env_var_at_pos_end() {
+        assert_eq!(
+            Env::var_at_pos(13, "hello ${world} and universe"),
+            "${world}"
+        );
     }
 }
